+++ ET/src/renderer/tr_local.h
@@ -417,7 +417,7 @@ typedef struct shader_s {
 	struct shader_s *currentShader;                     // current state if this is a state shader
 	struct shader_s *parentShader;                      // current state if this is a state shader
 	int currentState;                                   // current state index for cycle purposes
-	long expireTime;                                    // time in milliseconds this expires
+	int expireTime;                                     // time in milliseconds this expires
 
 	struct shader_s *remappedShader;                    // current shader this one is remapped too
